requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.7
cssselect2==0.8.0
dj-database-url==3.0.0
Django==5.2.1
djangorestframework==3.16.0
et_xmlfile==2.0.0
fonttools==4.59.2
html5lib==1.1
idna==3.10
lxml==6.0.1
numpy==2.2.5
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.3
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.30.0
pyhanko-certvalidator==0.28.0
pypdf==6.0.0
pyphen==0.17.2
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.3
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.9.0
xhtml2pdf==0.2.17
zopfli==0.2.3.post1
